+++ .github/scripts/release.py
@@ -15,4 +15,3 @@
 # * Git tag with version
 # * Git push branch and tag
 # * Cargo publish in all workspace members
-