+++ src/support/mod.rs
@@ -1,19 +1,22 @@
+//! Support for external crates.
+
+mod alloy_rlp;
 mod arbitrary;
 mod ark_ff;
 mod ark_ff_04;
 mod bn_rs;
 mod fastrlp;
 mod num_bigint;
-mod postgres;
+pub mod postgres;
 mod primitive_types;
 mod proptest;
 mod pyo3;
 mod quickcheck;
 mod rand;
 mod rlp;
-mod scale;
+pub mod scale;
 mod serde;
-mod sqlx;
+pub mod sqlx;
 mod valuable;
 mod zeroize;