+++ src/support/quickcheck.rs
@@ -1,6 +1,7 @@
 //! Support for the [`quickcheck`](https://crates.io/crates/quickcheck) crate.
+
 #![cfg(feature = "quickcheck")]
-#![cfg_attr(has_doc_cfg, doc(cfg(feature = "quickcheck")))]
+#![cfg_attr(docsrs, doc(cfg(feature = "quickcheck")))]
 
 use crate::Uint;
 use quickcheck::{Arbitrary, Gen};
@@ -9,7 +10,7 @@ impl<const BITS: usize, const LIMBS: usize> Arbitrary for Uint<BITS, LIMBS> {
     fn arbitrary(g: &mut Gen) -> Self {
         let mut limbs = [0; LIMBS];
         if let Some((last, rest)) = limbs.split_last_mut() {
-            for limb in rest.iter_mut() {
+            for limb in rest {
                 *limb = u64::arbitrary(g);
             }
             *last = u64::arbitrary(g) & Self::MASK;