+++ src/support/serde.rs
@@ -1,6 +1,7 @@
 //! Support for the [`serde`](https://crates.io/crates/serde) crate.
+
 #![cfg(feature = "serde")]
-#![cfg_attr(has_doc_cfg, doc(cfg(feature = "serde")))]
+#![cfg_attr(docsrs, doc(cfg(feature = "serde")))]
 
 use crate::{nbytes, Bits, Uint};
 use core::fmt::{Formatter, Result as FmtResult};
@@ -10,28 +11,69 @@ use serde::{
 };
 use std::{fmt::Write, str};
 
+/// Canonical serialization for all human-readable instances of `Uint<0, 0>`,
+/// and minimal human-readable `Uint<BITS, LIMBS>::ZERO` for any bit size.
+const ZERO_STR: &str = "0x0";
+
+impl<const BITS: usize, const LIMBS: usize> Uint<BITS, LIMBS> {
+    fn serialize_human_full<S: Serializer>(&self, s: S) -> Result<S::Ok, S::Error> {
+        if BITS == 0 {
+            return s.serialize_str(ZERO_STR);
+        }
+
+        let mut result = String::with_capacity(2 + nbytes(BITS) * 2);
+        result.push_str("0x");
+
+        self.as_le_bytes()
+            .iter()
+            .rev()
+            .try_for_each(|byte| write!(result, "{byte:02x}"))
+            .unwrap();
+
+        s.serialize_str(&result)
+    }
+
+    fn serialize_human_minimal<S: Serializer>(&self, s: S) -> Result<S::Ok, S::Error> {
+        if BITS == 0 {
+            return s.serialize_str(ZERO_STR);
+        }
+
+        let le_bytes = self.as_le_bytes();
+        let mut bytes = le_bytes.iter().rev().skip_while(|b| **b == 0);
+
+        // We avoid String allocation if there is no non-0 byte
+        // If there is a first byte, we allocate a string, and write the prefix
+        // and first byte to it
+        let mut result = match bytes.next() {
+            Some(b) => {
+                let mut result = String::with_capacity(2 + nbytes(BITS) * 2);
+                write!(result, "0x{b:x}").unwrap();
+                result
+            }
+            None => return s.serialize_str(ZERO_STR),
+        };
+        bytes
+            .try_for_each(|byte| write!(result, "{byte:02x}"))
+            .unwrap();
+
+        s.serialize_str(&result)
+    }
+
+    fn serialize_binary<S: Serializer>(&self, s: S) -> Result<S::Ok, S::Error> {
+        s.serialize_bytes(&self.to_be_bytes_vec())
+    }
+}
+
 /// Serialize a [`Uint`] value.
 ///
 /// For human readable formats a `0x` prefixed lower case hex string is used.
 /// For binary formats a byte array is used. Leading zeros are included.
 impl<const BITS: usize, const LIMBS: usize> Serialize for Uint<BITS, LIMBS> {
     fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-        let bytes = self.to_be_bytes_vec();
         if serializer.is_human_readable() {
-            // Special case for zero, which encodes as `0x0`.
-            if BITS == 0 {
-                return serializer.serialize_str("0x0");
-            }
-            // OPT: Allocation free method.
-            let mut result = String::with_capacity(2 * Self::BYTES + 2);
-            result.push_str("0x");
-            for byte in bytes {
-                write!(result, "{byte:02x}").unwrap();
-            }
-            serializer.serialize_str(&result)
+            self.serialize_human_minimal(serializer)
         } else {
-            // Write as bytes directly
-            serializer.serialize_bytes(&bytes[..])
+            self.serialize_binary(serializer)
         }
     }
 }
@@ -42,7 +84,7 @@ impl<const BITS: usize, const LIMBS: usize> Serialize for Uint<BITS, LIMBS> {
 impl<'de, const BITS: usize, const LIMBS: usize> Deserialize<'de> for Uint<BITS, LIMBS> {
     fn deserialize<D: Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
         if deserializer.is_human_readable() {
-            deserializer.deserialize_str(StrVisitor)
+            deserializer.deserialize_any(HrVisitor)
         } else {
             deserializer.deserialize_bytes(ByteVisitor)
         }
@@ -51,7 +93,11 @@ impl<'de, const BITS: usize, const LIMBS: usize> Deserialize<'de> for Uint<BITS,
 
 impl<const BITS: usize, const LIMBS: usize> Serialize for Bits<BITS, LIMBS> {
     fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-        self.as_uint().serialize(serializer)
+        if serializer.is_human_readable() {
+            self.as_uint().serialize_human_full(serializer)
+        } else {
+            self.as_uint().serialize_binary(serializer)
+        }
     }
 }
 
@@ -61,54 +107,41 @@ impl<'de, const BITS: usize, const LIMBS: usize> Deserialize<'de> for Bits<BITS,
     }
 }
 
-/// Serde Visitor for human readable formats
-struct StrVisitor<const BITS: usize, const LIMBS: usize>;
+/// Serde Visitor for human readable formats.
+///
+/// Accepts either a primitive number, a decimal or a hexadecimal string.
+struct HrVisitor<const BITS: usize, const LIMBS: usize>;
 
-impl<'de, const BITS: usize, const LIMBS: usize> Visitor<'de> for StrVisitor<BITS, LIMBS> {
+impl<'de, const BITS: usize, const LIMBS: usize> Visitor<'de> for HrVisitor<BITS, LIMBS> {
     type Value = Uint<BITS, LIMBS>;
 
     fn expecting(&self, formatter: &mut Formatter) -> FmtResult {
         write!(formatter, "a {} byte hex string", nbytes(BITS))
     }
 
-    fn visit_str<E>(self, value: &str) -> Result<Self::Value, E>
-    where
-        E: Error,
-    {
-        let value = trim_hex_prefix(value);
+    fn visit_u64<E: Error>(self, v: u64) -> Result<Self::Value, E> {
+        Uint::try_from(v).map_err(|_| Error::invalid_value(Unexpected::Unsigned(v), &self))
+    }
 
-        // ensure the string is the correct length (two characters in the string per
-        // byte) exception: zero, for a Uint where BITS == 0
-        if BITS == 0 {
-            // special case, this class of ints has one member, zero.
-            // zero is represented as "0x0" only
-            if value != "0" {
-                return Err(Error::invalid_value(Unexpected::Str(value), &self));
-            }
+    fn visit_u128<E: Error>(self, v: u128) -> Result<Self::Value, E> {
+        // `Unexpected::Unsigned` cannot contain a `u128`
+        Uint::try_from(v).map_err(Error::custom)
+    }
+
+    fn visit_str<E: Error>(self, value: &str) -> Result<Self::Value, E> {
+        // Shortcut for common case
+        if value == ZERO_STR {
             return Ok(Uint::<BITS, LIMBS>::ZERO);
         }
-        if nbytes(BITS) * 2 < value.len() {
-            return Err(Error::invalid_length(value.len(), &self));
-        }
-
-        let mut limbs = [0; LIMBS];
-        for (i, chunk) in value.as_bytes().rchunks(16).enumerate() {
-            let chunk = str::from_utf8(chunk)
-                .map_err(|_| Error::invalid_value(Unexpected::Str(value), &self))?;
-            let limb = u64::from_str_radix(chunk, 16)
-                .map_err(|_| Error::invalid_value(Unexpected::Str(value), &self))?;
-            if limb == 0 {
-                continue;
-            }
-            if i >= LIMBS {
-                return Err(Error::invalid_value(Unexpected::Str(value), &self));
-            }
-            limbs[i] = limb;
-        }
-        if BITS > 0 && limbs[LIMBS - 1] > Self::Value::MASK {
+        // `ZERO_STR` is the only valid serialization of `Uint<0, 0>`, so if we
+        // have not shortcut, we are in an error case
+        if BITS == 0 {
             return Err(Error::invalid_value(Unexpected::Str(value), &self));
         }
-        Ok(Uint::from_limbs(limbs))
+
+        value
+            .parse()
+            .map_err(|_| Error::invalid_value(Unexpected::Str(value), &self))
     }
 }
 
@@ -122,29 +155,16 @@ impl<'de, const BITS: usize, const LIMBS: usize> Visitor<'de> for ByteVisitor<BI
         write!(formatter, "{BITS} bits of binary data in big endian order")
     }
 
-    fn visit_bytes<E>(self, value: &[u8]) -> Result<Self::Value, E>
-    where
-        E: Error,
-    {
+    fn visit_bytes<E: Error>(self, value: &[u8]) -> Result<Self::Value, E> {
         if value.len() != nbytes(BITS) {
-            return Err(E::invalid_length(value.len(), &self));
+            return Err(Error::invalid_length(value.len(), &self));
         }
         Uint::try_from_be_slice(value).ok_or_else(|| {
-            E::invalid_value(
-                Unexpected::Other(&format!("Value to large for Uint<{BITS}>")),
+            Error::invalid_value(
+                Unexpected::Other(&format!("too large for Uint<{BITS}>")),
                 &self,
             )
         })
-    }
-}
-
-/// Helper function to remove optionally `0x` prefix from hex strings.
-#[allow(clippy::missing_const_for_fn)]
-fn trim_hex_prefix(str: &str) -> &str {
-    if str.len() >= 2 && (&str[..2] == "0x" || &str[..2] == "0X") {
-        &str[2..]
-    } else {
-        str
     }
 }
 
@@ -163,7 +183,37 @@ mod tests {
                 let deserialized = serde_json::from_str(&serialized).unwrap();
                 assert_eq!(value, deserialized);
             });
+            proptest!(|(value: Bits<BITS, LIMBS>)| {
+                let serialized = serde_json::to_string(&value).unwrap();
+                let deserialized = serde_json::from_str(&serialized).unwrap();
+                assert_eq!(value, deserialized);
+            });
         });
+    }
+
+    #[test]
+    fn test_human_readable_de() {
+        let jason = r#"[
+            1,
+            "0x1",
+            "0o1",
+            "0b1"
+        ]"#;
+        let numbers: Vec<Uint<1, 1>> = serde_json::from_str(jason).unwrap();
+        ruint_macro::uint! {
+            assert_eq!(numbers, vec![1_U1, 1_U1, 1_U1, 1_U1]);
+        }
+
+        let jason = r#"[
+            "",
+            "0x",
+            "0o",
+            "0b"
+        ]"#;
+        let numbers: Vec<Uint<1, 1>> = serde_json::from_str(jason).unwrap();
+        ruint_macro::uint! {
+            assert_eq!(numbers, vec![0_U1, 0_U1, 0_U1, 0_U1]);
+        }
     }
 
     #[test]
@@ -202,32 +252,7 @@ mod tests {
             serialized.push('0');
             serialized.push('"');
             let deserialized = serde_json::from_str::<Uint<BITS, LIMBS>>(&serialized);
-            assert!(deserialized.is_err());
-        });
-    }
-
-    #[test]
-    fn test_serde_zero_invalid_size_error() {
-        // Test that if we add a zero to a large zero string, we get an error.
-        // This is done by replacing a max string `0xffff...` with zeros: `0x0000...`
-        const_for!(BITS in SIZES {
-            const LIMBS: usize = nlimbs(BITS);
-            let value = Uint::<BITS, LIMBS>::MAX;
-            let mut serialized = serde_json::to_string(&value).unwrap();
-
-            // replace `f`s with `0`s
-            serialized = serialized.replace('f', "0");
-            // ensure format of serialized value is correct ("0x...")
-            assert_eq!(&serialized[..3], "\"0x");
-            // last character should be a quote
-            assert_eq!(&serialized[serialized.len() - 1..], "\"");
-
-            // strip the last character, add a zero, and finish with a quote
-            serialized.pop();
-            serialized.push('0');
-            serialized.push('"');
-            let deserialized = serde_json::from_str::<Uint<BITS, LIMBS>>(&serialized);
-            assert!(deserialized.is_err());
+            assert!(deserialized.is_err(), "{BITS} {serialized}");
         });
     }
 }