+++ CHANGELOG.md
@@ -9,14 +9,31 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Added 
-
 - Introduce `ark-ff-04` feature flag for conversion to `ark-ff@0.4` types.
+- Support for [`alloy-rlp`](https://github.com/alloy-rs/rlp)
+- MSRV (Minimum Supported Rust Version) is now set at 1.65.0, from previously undefined.
 
 ### Changed
 
-- Updated `fastrlp` to `0.3` and `pyo3` to `0.18`.
+- Make `serde::Deserialize` impl more permissive
+- Use Ethereum `Quantity` encoding for serde serialization when human-readable
+- Fix error in `from_base_be` that allowed instantiation of overflowing `Uint`.
+- Updated `fastrlp` to `0.3`, `pyo3` to `0.19`, and `sqlx-core` to `0.7`.
+- Improved `fastrlp` perfomance
+- Improved `proptest` performance
+- Made `support` module and its modules public
+- Made more `algorithm` functions public
+- Constified `as_le_slice` and `as_le_bytes`
 
+### Removed
+
+- Automatic detection of nightly features. Enable them instead with the `nightly` cargo feature.
+- Dependency on `derive_more`
+
+### Fixed
+
+- `from_base_le` implementation by reversing the input iterator
+>>>>>>> main:CHANGELOG.md
 ## [1.8.0] — 2023-04-19
 
 ### Added